generic_requirements.txt
anyio==3.6.1
asgiref==3.5.2
black==22.6.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
Django==4.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
fastapi==0.80.0
greenlet==1.1.3
h11==0.13.0
httptools==0.4.0
idna==3.3
mypy-extensions==0.4.3
pathspec==0.9.0
pika==1.3.0
platformdirs==2.5.2
pydantic==1.9.2
PyJWT==2.4.0
python-dotenv==0.20.0
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
sniffio==1.2.0
SQLAlchemy==1.4.40
sqlparse==0.4.2
starlette==0.19.1
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
uvloop==0.16.0
watchfiles==0.16.1
websockets==10.3
